+++ app/inputs/search_select_filter_input.rb
@@ -7,6 +7,20 @@ class SearchSelectFilterInput < SearchSelectInput
   end
 
   def input_method
-    eq_input_name
+    "#{input_name}_eq"
+  end
+
+  def input_html_options_name
+    "#{object_name}[#{input_method}]"
+  end
+
+  def input_value
+    result = valid_object.conditions.find do |condition|
+      condition.attributes.map(&:name).include?(input_name.to_s)
+    end
+
+    return unless result
+
+    result.values.first.value
   end
 end