+++ lib/activeadmin_addons/support/select_input_base.rb
@@ -3,7 +3,7 @@ require_relative "input_helpers/input_methods"
 require_relative "input_helpers/input_html_helpers"
 
 module ActiveAdminAddons
-  class InputBase < Formtastic::Inputs::StringInput
+  class SelectInputBase < Formtastic::Inputs::SelectInput
     include InputOptionsHandler
     include InputMethods
     include InputHtmlHelpers